+++ dcm_common/db/__init__.py
@@ -11,21 +11,29 @@ from .key_value_store.adapter.http import HTTPKeyValueStoreAdapter
 from .key_value_store.adapter.postgres import PostgreSQLAdapter14
 
 
-# FIXME fix naming conflict for PostgreSQLAdapter14 in key_value_store and sqlimport subprocess
+# FIXME fix naming conflict for PostgreSQLAdapter14 in key_value_store and
+# sqlimport subprocess
 
 
 def check_psycopg_dependencies():
     """
     Checks whether the package `psycopg` can be imported without segfault.
     """
+    # pylint: disable=import-outside-toplevel
     import sys
     import subprocess
 
     try:
-        subprocess.run([sys.executable, "-c", "import psycopg"], check=True)
+        subprocess.run(
+            [sys.executable, "-c", "import psycopg"],
+            check=True,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        )
     except subprocess.CalledProcessError as exc_info:
         return False, str(exc_info)
     return True, ""
+
 
 # safeguard imports
 if (psycopg_ok := check_psycopg_dependencies())[0]: